+++ src/main/java/com/mobilecompany/entities/Contract.java
@@ -18,7 +18,7 @@ public class Contract {
     private BigDecimal balance;
 
     @Column(name = "isBlocked")
-    private byte isBlocked;
+    private int isBlocked;
 
     public Contract() {
     }
@@ -51,10 +51,10 @@ public class Contract {
         this.balance = balance;
     }
 
-    public byte getIsBlocked() {
+    public int getIsBlocked() {
         return isBlocked;
     }
-    public void setIsBlocked(byte isBlocked) {
+    public void setIsBlocked(int isBlocked) {
         this.isBlocked = isBlocked;
     }
